feat: 支持运行时配置更新与拨号前清理残留连接
- manager.go: 新增 SetProfileIndex/SetAPN 方法支持运行时更新拨号配置

- manager.go: doConnect 中 StartNetworkInterface 前调用 StopAnyNetworkInterface 清理残留数据呼叫

- qrtr_transport_linux.go: 修正 QRTR 传输层参数

+++ pkg/qmi/qrtr_transport_linux.go
@@ -178,7 +178,7 @@ func (t *qrtrTransport) Write(p []byte) (int, error) {
 	}
 	body := frame[fh.headerSize:]
 
-	if fh.serviceType == ServiceControl {
+	if fh.serviceType == uint16(ServiceControl) {
 		if err := t.handleCTLWrite(body); err != nil {
 			return 0, err
 		}
@@ -571,7 +571,7 @@ func marshalCTLResponseFrame(txID uint8, msgID uint16, tlvs []TLV) []byte {
 		Length:        uint16(len(tlvBytes)),
 	}
 	body := append(ctlH.Marshal(), tlvBytes...)
-	return append(marshalFrameHeader(ServiceControl, 0, len(body)), body...)
+	return append(marshalFrameHeader(uint16(ServiceControl), 0, len(body)), body...)
 }
 
 func (t *qrtrTransport) replyCTLError(txID uint8, msgID uint16, errCode uint16) error {
